[RENAME] #8 - data class 패키지 수정
