+++ app/src/main/java/co/kr/woowahan_repo/data/model/request/OAuthAccessTokenRequest.kt
@@ -1,4 +1,4 @@
-package co.kr.woowahan_repo.data.model
+package co.kr.woowahan_repo.data.model.request
 
 import com.google.gson.annotations.SerializedName
