+++ app/src/main/java/co/kr/woowahan_repo/presentation/viewmodel/SignInViewModel.kt
@@ -3,7 +3,7 @@ package co.kr.woowahan_repo.presentation.viewmodel
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import co.kr.woowahan_repo.BuildConfig
-import co.kr.woowahan_repo.data.model.OAuthAccessTokenRequest
+import co.kr.woowahan_repo.data.model.request.OAuthAccessTokenRequest
 import co.kr.woowahan_repo.di.ServiceLocator
 import kotlinx.coroutines.launch
 import timber.log.Timber
